% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHT.R
\name{FHT}
\alias{FHT}
\alias{dFHT}
\alias{pFHT}
\alias{qFHT}
\alias{rFHT}
\title{The First Hitting Time Distribution}
\usage{
dFHT(t, x0, nu, kappa, sigma)

pFHT(t, x0, nu, kappa, sigma)

qFHT(p, x0, nu, kappa, sigma)

rFHT(n, x0, nu, kappa, sigma, k = 3, qTail = 0.95)
}
\arguments{
\item{t}{random values; vector}

\item{x0}{starting point; scalar}

\item{nu}{lower barrier; scalar}

\item{kappa}{upper reflection barrier; scalar}

\item{sigma}{volatility; scalar}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{k}{number of pieces to be included in proposal, default k = 3}

\item{qTail}{the user-defined qth quantile for the tail, default q = 0.95}
}
\description{
Density, distribution, quantile, and random number generation function
}
\examples{
dFHT(1, 10, 3.9, 20, 3)
dFHT(c(1:5), 10, 3.9, 20, 3)
pFHT(1, 10, 3.9, 20, 3)
pFHT(c(1:5), 10, 3.9, 20, 3)
}
