% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHTRBM.R
\name{FHTRBM}
\alias{FHTRBM}
\alias{dfhtrbm}
\alias{pfhtrbm}
\alias{qfhtrbm}
\alias{rfhtrbm}
\title{The First Hitting Time Distribution}
\usage{
dfhtrbm(x, x0, nu, kappa, sigma, log = FALSE)

pfhtrbm(q, x0, nu, kappa, sigma, lower.tail = TRUE, log.p = FALSE)

qfhtrbm(p, x0, nu, kappa, sigma, lower.tail = TRUE, log.p = FALSE)

rfhtrbm(n, x0, nu, kappa, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{x0}{scalar; starting point of the reflected Brownian motion.}

\item{nu}{scalar; lower boundary of the reflected Brownian motion.}

\item{kappa}{scalar; upper reflection barrier of the reflected Brownian
motion.}

\item{sigma}{scalar; volatility of the reflected Brownian motion.}

\item{log, log.p}{logical; default is FALSE; if TRUE, probabilities p
are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities
are \eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\description{
Density, distribution, quantile, and random number
    generation function of first hitting time (hitting down) distribution
    of reflected Brownian motion.
}
\examples{
dfhtrbm(1, 10, 3.9, 20, 3)
dfhtrbm(c(1:5), 10, 3.9, 20, 3)
pfhtrbm(1, 10, 3.9, 20, 3)
pfhtrbm(c(1:5), 10, 3.9, 20, 3)
}
\references{
Hu, Q., Wang, Y., and Yang, X. (2012). The hitting time density
for a reflected Brownian motion.
\emph{Computational Economics}, \bold{40}, 1–18.
}
