% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFhtrbm.R
\name{fitFhtrbm}
\alias{fitFhtrbm}
\title{Recurrent Events Threshold Regression}
\usage{
fitFhtrbm(
  formula,
  data,
  x0 = 10,
  nu = 3.9,
  frailty = c("correlated", "independent", "shared", "none"),
  initial = NULL,
  thin = 10,
  nburnin = 3000,
  niter = 6000,
  nchains = 1
)
}
\arguments{
\item{formula}{A formula object, with the response on the left
of a "~" operator, and the predictors on the right. The response must be
a recurrent event object as returned by function Recur.}

\item{data}{A data frame includes all the variables in "formula".
Data frame should not include intercept.}

\item{x0}{The starting point of Brownian motion, which is a known
parameter in model.}

\item{nu}{The lower boundary of Brownian motion, which is a known
parameter in model.}

\item{frailty}{Frailty model type, \code{"correlated"}, \code{"independent"},
\code{"shared"}, and \code{"none"}.}

\item{initial}{The initial value for MCMC, which is optional.
Default is Null.}

\item{thin}{Thinning number of MCMC chain.}

\item{nburnin}{Burn-in number of MCMC chain.}

\item{niter}{Total number of iteration of MCMC chain before thin
and burn-in.}

\item{nchains}{Number of chains}
}
\description{
The function fitFhtrbm return MCMC samples.
}
\examples{
data(simuEvtDat)
# need to run longer MCMC chain
# note that simuEvtDat dataset does not include intercept column
\dontrun{
mcmc <- fitFhtrbm(Recur(time = time, id = id, event = event) ~ x1 + x2|x1 + x2,
                 data = simuEvtDat,
                 frailty = "correlated",
                 thin = 10, nburnin = 3000, niter = 10000, nchains = 1)
}
}
