% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simCovData}
\alias{simCovData}
\title{Simulated Covariate Data Frame}
\format{
A data frame with 200 rows and 4 variables.
}
\description{
A simulated data frame for covariates named
\code{id}, \code{censorTime}, \code{x1}, and \code{x2}, where
\itemize{
    \item \code{id}: Subjects identification;
    \item \code{censorTime}: Subject follow-up time;
    \item \code{x1}: Continuous variable.
    \item \code{x2}: Continuous variable.
}
}
\details{
The sample dataset contains the subject level covariates generated from 
the fitted Copula model.
}
