% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPath.R
\name{simPath}
\alias{simPath}
\title{Reflected Brownian Motion Path}
\usage{
simPath(x0, nu, kappa, sigma, time_idx, plot_path = FALSE)
}
\arguments{
\item{x0}{The starting point of reflected Brownian motion, which is a scalar.}

\item{nu}{The lower boundary of reflected Brownian motion, which is a scalar.}

\item{kappa}{The upper reflection barrier of reflected Brownian motion,
which is a scalar.}

\item{sigma}{The volatility of reflected Brownian motion, which is a scalar.}

\item{time_idx}{The discrete time index, which is a vector}

\item{plot_path}{Plot the path if TRUE. Default is FALSE.}
}
\value{
A matrix contains three column, time_idx, Xt, proc_idx. The column
   proc_idx indicates which process the Brownian motion belongs to.
}
\description{
Simulating sample path of reflected Brownian motion.
    The reflected Brownian motion begins at starting point. The process
    stops when it reach the lower boundary and then reset the process to
    be the value of the starting point.
}
\examples{
set.seed(1)
time_idx <- seq(0, 200, 1)
sim_list <- simPath(x0 = 10, nu = 3.9, kappa = 20, sigma = 1.5,
                    time_idx, plot_path = TRUE)
}
