% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuCovDat}
\alias{simuCovDat}
\title{Simulated Covariate Data Set}
\format{
A data frame with 400 rows and 4 variables.
}
\description{
A simulated data set for variables named
\code{id}, \code{censorTime}, \code{x1}, and \code{x2}, where
\itemize{
    \item \code{id}: Subjects identification;
    \item \code{censorTime}: Subject follow-up time;
    \item \code{x1}: Continuous covariate x1;
    \item \code{x2}: Continuous covariate x2;
}
}
\details{
The sample dataset contains the subject-level covariates generated from
the fitted Copula model.
}
