% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuEvtDat}
\alias{simuEvtDat}
\title{Simulated Event Time Data Set}
\format{
A data frame with 3779 rows and 5 variables.
}
\description{
The simulated event time data set contains the calender time of the
recurrent events, the event indicator and the subject covariates.
\code{id}, \code{event}, \code{time}, \code{x1} and \code{x2}, where
\itemize{
    \item \code{id}: The patient id;
    \item \code{event}: Event indicator, '1' for event, and '0' for censored;
    \item \code{time}: Event or censored time (day);
    \item \code{x1}; Continuous covariate x1;
    \item \code{x2}; Continuous covariate x2;
}
}
