% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHTRBM.R
\name{FHTRBM}
\alias{FHTRBM}
\alias{dFHTRBM}
\alias{pFHTRBM}
\alias{qFHTRBM}
\alias{rFHTRBM}
\title{The First Hitting Time Distribution}
\usage{
dFHTRBM(x, x0, nu, kappa, sigma, log = FALSE)

pFHTRBM(x, x0, nu, kappa, sigma)

qFHTRBM(p, x0, nu, kappa, sigma)

rFHTRBM(n, x0, nu, kappa, sigma, k = 3, qTail = 0.95)
}
\arguments{
\item{x}{random values; vector}

\item{x0}{starting point; scalar}

\item{nu}{lower barrier; scalar}

\item{kappa}{upper reflection barrier; scalar}

\item{sigma}{volatility; scalar}

\item{log}{default is FALSE}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{k}{number of pieces to be included in proposal, default k = 3}

\item{qTail}{the user-defined qth quantile for the tail, default q = 0.95}
}
\description{
Density, distribution, quantile, and random number
    generation function
}
\examples{
dFHTRBM(1, 10, 3.9, 20, 3)
dFHTRBM(c(1:5), 10, 3.9, 20, 3)

pFHTRBM(1, 10, 3.9, 20, 3)
pFHTRBM(c(1:5), 10, 3.9, 20, 3)
}
